requirements.txt
autopep8==1.5.7
mysql-connector-python==8.0.26
protobuf==3.18.0
pycodestyle==2.7.0
toml==0.10.2
